refactor(components/atom/tooltip): api tab
